requirement.txt
Django==1.10
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-tinymce==2.3.0
markdown2==2.3.1
Pillow==3.3.0
